refactor: move constants and utilities from src to cypress folder
