+++ cypress/constants/page-titles.ts
