+++ cypress/constants/styles.ts
