+++ cypress/e2e/is-available.spec.cy.js
@@ -1,4 +1,4 @@
-import { CHECK } from '../../src/constants/test-names';
+import { CHECK } from '../constants/test-names';
 
 describe(`${CHECK.APP_IS_AVAILABLE}`, () => {
   it(`${CHECK.LOCALHOST_AVAILABLE}`, () => {
