+++ cypress/utils/check-circle-props.ts
