+++ cypress/utils/check-circles-length.ts
