+++ cypress/utils/get-circle-props.ts
