+++ tsconfig.json
@@ -16,8 +16,18 @@
     "noEmit": true,
     "jsx": "react-jsx",
     "paths": {
-      "react":["./node_modules/@types/react"]
+      "react": ["./node_modules/@types/react"]
     }
   },
-  "include": ["src"]
+  "include": [
+    "src",
+    "cypress/constants/test-names.ts",
+    "cypress/utils/get-circle-props.ts",
+    "cypress/utils/check-circles-length.ts",
+    "cypress/utils/check-circle-props.ts",
+    "src/constants/mock-data.ts",
+    "cypress/constants/test-selectors.ts",
+    "cypress/constants/page-titles.ts",
+    "cypress/constants/styles.ts"
+  ]
 }
